Refactor(BlackDuckAccumulator): Move accumulator class to package of same name
