+++ src/test/java/com/synopsys/integration/alert/provider/blackduck/task/accumulator/BlackDuckAccumulatorTest.java
@@ -1,4 +1,4 @@
-package com.synopsys.integration.alert.provider.blackduck.task;
+package com.synopsys.integration.alert.provider.blackduck.task.accumulator;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -30,7 +30,6 @@ import com.synopsys.integration.alert.database.api.DefaultNotificationAccessor;
 import com.synopsys.integration.alert.descriptor.api.BlackDuckProviderKey;
 import com.synopsys.integration.alert.provider.blackduck.BlackDuckProperties;
 import com.synopsys.integration.alert.provider.blackduck.TestBlackDuckProperties;
-import com.synopsys.integration.alert.provider.blackduck.task.accumulator.BlackDuckAccumulatorSearchDateManager;
 import com.synopsys.integration.alert.provider.blackduck.validator.BlackDuckValidator;
 import com.synopsys.integration.alert.test.common.TestAlertProperties;
 import com.synopsys.integration.alert.test.common.TestProperties;
@@ -114,7 +113,6 @@ public class BlackDuckAccumulatorTest {
         BlackDuckAccumulator notificationAccumulator = createAccumulator(testBlackDuckProperties);
         BlackDuckAccumulator spiedAccumulator = Mockito.spy(notificationAccumulator);
         spiedAccumulator.accumulate();
-        assertTrue(providerTaskPropertiesAccessor.getTaskProperty(spiedAccumulator.getTaskName(), BlackDuckAccumulatorSearchDateManager.TASK_PROPERTY_KEY_LAST_SEARCH_END_DATE).isPresent());
         Mockito.verify(spiedAccumulator).retrieveAndStoreNotifications(Mockito.any());
     }
 
@@ -124,7 +122,6 @@ public class BlackDuckAccumulatorTest {
         BlackDuckAccumulator spiedAccumulator = Mockito.spy(notificationAccumulator);
         Mockito.when(spiedAccumulator.getMillisecondsToNextRun()).thenReturn(Optional.of(Long.MAX_VALUE));
         spiedAccumulator.accumulate();
-        assertTrue(providerTaskPropertiesAccessor.getTaskProperty(spiedAccumulator.getTaskName(), BlackDuckAccumulatorSearchDateManager.TASK_PROPERTY_KEY_LAST_SEARCH_END_DATE).isPresent());
         Mockito.verify(spiedAccumulator).retrieveAndStoreNotifications(Mockito.any());
     }
